chore: move methods from cartola file to preprocessing pipeline
